% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CreateRandomSwissPlot}
\alias{CreateRandomSwissPlot}
\title{Create Random Swiss Plot}
\usage{
CreateRandomSwissPlot(filePath)
}
\arguments{
\item{filePath}{The full qualified path at which to write the plot}
}
\description{
Uses the other functions in the swiptapi package to create a plot of busses in Switzerland.
}
