% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{SwiPtApiClient}
\alias{SwiPtApiClient}
\title{Object to interact with the Swiss Public Transit API}
\description{
An object to streamline interactions with the Swiss Public Transit API.
}
\references{
(https://transport.opendata.ch/)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SwiPtApiClient$new()}}
\item \href{#method-SearchAroundPoint}{\code{SwiPtApiClient$SearchAroundPoint()}}
\item \href{#method-clone}{\code{SwiPtApiClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a swiptapi API client
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SwiPtApiClient$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Returns a initialized API client
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SearchAroundPoint"></a>}}
\if{latex}{\out{\hypertarget{method-SearchAroundPoint}{}}}
\subsection{Method \code{SearchAroundPoint()}}{
Searches for Points of Interest around submitted lat/long pair
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SwiPtApiClient$SearchAroundPoint(lati, longi)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lati}}{The latitude around which to search for points of interest}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns a cleaned dictionary of points of interest
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SwiPtApiClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
