% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{GetRandomSwissPoint}
\alias{GetRandomSwissPoint}
\title{Get Random Swiss Point}
\usage{
GetRandomSwissPoint()
}
\value{
a named list of x and y lat/long coordinates
}
\description{
Uses basic math to pick a point in a circle that roughly approximates Switzerland
}
