% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\docType{data}
\name{SwiPtApiClient}
\alias{SwiPtApiClient}
\title{Object to interact with the Swiss Public Transit API}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SwiPtApiClient
}
\description{
An object to streamline interactions with the Swiss Public Transit API.
}
\references{
(https://transport.opendata.ch/)
}
\keyword{datasets}
